PMM-10059 refactor
