+++ public/app/percona/integrated-alerting/components/Table/Filter/Filter.constants.ts
@@ -1,0 +1,2 @@
+export const SEARCH_SELECT_FIELD_NAME = 'search-select';
+export const SEARCH_INPUT_FIELD_NAME = 'search-text-input';
